move segments to class
