+++ packages/backend/src/modules/finality/analyzers/arbitrum/ArbitrumT2IAnalyzer.ts
@@ -1,18 +1,22 @@
 import { Logger } from '@l2beat/backend-tools'
 import { assert, ProjectId, TrackedTxsConfigSubtype } from '@l2beat/shared-pure'
 
+import zlib from 'zlib'
 import { Database } from '@l2beat/database'
-import { BlobProvider, EVMTransaction, RpcClient } from '@l2beat/shared'
+import { Blob, BlobProvider, EVMTransaction, RpcClient } from '@l2beat/shared'
 import { utils } from 'ethers'
-import { RlpSerializable } from '../../utils/rlpDecode'
+import { byteArrFromHexStr } from '../../utils/byteArrFromHexStr'
+import { RlpSerializable, rlpDecode } from '../../utils/rlpDecode'
 import { BaseAnalyzer } from '../types/BaseAnalyzer'
 import type { L2Block, Transaction } from '../types/BaseAnalyzer'
+import { blobsToData } from './blobsToData'
 import { calculateDelaysFromSegments } from './calculateDelaysFromSegments'
-import { getSegmentsFromBlobs, getSegmentsFromCalldata } from './getSegments'
+import { numberToByteArr } from './utils'
 
 const calldataSignature =
   'addSequencerL2BatchFromOrigin(uint256 sequenceNumber,bytes data,uint256 afterDelayedMessagesRead,address gasRefunder,uint256 prevMessageCount,uint256 newMessageCount)'
 const iface = new utils.Interface([`function ${calldataSignature}`])
+const BROTLI_COMPRESSION_TYPE = 0x00
 
 export class ArbitrumT2IAnalyzer extends BaseAnalyzer {
   constructor(
@@ -61,7 +65,7 @@ export class ArbitrumT2IAnalyzer extends BaseAnalyzer {
           tx.data,
         )
         assert(decodedInput[1], 'No data in calldata')
-        return getSegmentsFromCalldata(decodedInput[1])
+        return this.getSegmentsFromCalldata(decodedInput[1])
       }
       case 3: {
         assert(
@@ -77,10 +81,62 @@ export class ArbitrumT2IAnalyzer extends BaseAnalyzer {
             tx.blockNumber,
           )
         if (blobs.length === 0) return []
-        return getSegmentsFromBlobs(blobs)
+        return this.getSegmentsFromBlobs(blobs)
       }
       default:
         throw new Error(`Unsupported transaction type: ${tx.type}`)
     }
+  }
+
+  getSegmentsFromBlobs(relevantBlobs: Blob[]): RlpSerializable[] {
+    const blobs = relevantBlobs.map(({ data }) => byteArrFromHexStr(data))
+    const payload = blobsToData(blobs)
+    const decompressed = this.decompressPayload(payload)
+    // I do not understand why this is necessary, but it is
+    // My guess is that the shape of the data is always the same,
+    // so it's an optimization technique to save couple of bytes
+    const rlpEncoded = this.concatWithLengthForRlp(decompressed)
+    const segments = rlpDecode(rlpEncoded)
+    assert(Array.isArray(segments), 'Expected segments to be an array')
+
+    return segments
+  }
+
+  getSegmentsFromCalldata(data: string): RlpSerializable[] {
+    const payload = byteArrFromHexStr(data)
+    const decompressed = this.decompressPayload(payload)
+    // I do not understand why this is necessary, but it is
+    // My guess is that the shape of the data is always the same,
+    // so it's an optimization technique to save couple of bytes
+    const rlpEncoded = this.concatWithLengthForRlp(decompressed)
+    const segments = rlpDecode(rlpEncoded)
+    assert(Array.isArray(segments), 'Expected segments to be an array')
+
+    return segments
+  }
+
+  decompressPayload(payload: Uint8Array): Uint8Array {
+    const compressionType = payload[0]
+    assert(
+      compressionType === BROTLI_COMPRESSION_TYPE,
+      `Expected compression type to be ${BROTLI_COMPRESSION_TYPE}, got ${compressionType}`,
+    )
+    const data = payload.slice(1)
+    const decompressedData = zlib.brotliDecompressSync(data)
+    return Uint8Array.from(decompressedData)
+  }
+
+  concatWithLengthForRlp(decompressedBytes: Uint8Array) {
+    const totalLength = decompressedBytes.length
+    const lengthBytes = numberToByteArr(totalLength)
+    const lengthBytesLength = lengthBytes.length
+    const lengthByte = 0xf7 + lengthBytesLength
+    const lengthByteBytes = numberToByteArr(lengthByte)
+    const concatenatedWithLength = new Uint8Array([
+      ...lengthByteBytes,
+      ...lengthBytes,
+      ...decompressedBytes,
+    ])
+    return concatenatedWithLength
   }
 }
